+++ packages/svelte/tests/hydration/samples/custom-element-with-settable-only-property/main.svelte
@@ -1,5 +1,14 @@
-<script>
-  import "./custom-element-with-settable-only-property.js"
+<script module>
+	customElements.define("custom-element-with-settable-only-property", class CustomElement extends HTMLElement {
+		set prop(n) {
+			this.value = n;
+		}
+
+		get prop() {
+			// invoking this getter shouldn't make hydration fail
+			throw new Error("This value is not gettable");
+		}
+	});
 </script>
 
 <custom-element-with-settable-only-property prop="8"></custom-element-with-settable-only-property>
